Refactor setup controller and helper functions controller and add powerpack features in plugins contoller
