+++ includes/class-wc-calypso-bridge-plugins.php
@@ -91,6 +91,51 @@ class WC_Calypso_Bridge_Plugins {
 		add_action( 'admin_notices', array( $this, 'prevent_woocommerce_deactivation_notice' ), 10, 2 );
 		add_filter( 'manage_product_posts_columns', array( $this, 'remove_jetpack_stats_column' ), 100 );
 		add_filter( 'default_hidden_columns', array( $this, 'hide_product_columns' ), 100, 2 );
+
+		add_action( 'plugins_loaded', array( $this, 'disable_powerpack_features' ), 2 );
+	}
+
+	/**
+	 * Disables Specific Features within the Powerpack extension for Storefront.
+	 */
+	public function disable_powerpack_features() {
+		if ( ! class_exists( 'Storefront_Powerpack' ) ) {
+			return;
+		}
+
+		/**
+		 * List of Powerpack features able to disable
+		 *
+		 * 'storefront_powerpack_helpers_enabled'
+		 * 'storefront_powerpack_admin_enabled'
+		 * 'storefront_powerpack_frontend_enabled'
+		 * 'storefront_powerpack_customizer_enabled'
+		 * 'storefront_powerpack_header_enabled'
+		 * 'storefront_powerpack_footer_enabled'
+		 * 'storefront_powerpack_designer_enabled'
+		 * 'storefront_powerpack_layout_enabled'
+		 * 'storefront_powerpack_integrations_enabled'
+		 * 'storefront_powerpack_mega_menus_enabled'
+		 * 'storefront_powerpack_parallax_hero_enabled'
+		 * 'storefront_powerpack_checkout_enabled'
+		 * 'storefront_powerpack_homepage_enabled'
+		 * 'storefront_powerpack_messages_enabled'
+		 * 'storefront_powerpack_product_details_enabled'
+		 * 'storefront_powerpack_shop_enabled'
+		 * 'storefront_powerpack_pricing_tables_enabled'
+		 * 'storefront_powerpack_reviews_enabled'
+		 * 'storefront_powerpack_product_hero_enabled'
+		 * 'storefront_powerpack_blog_customizer_enabled'
+		 */
+		$disabled_powerpack_features = array(
+			'storefront_powerpack_designer_enabled',
+			'storefront_powerpack_mega_menus_enabled',
+			'storefront_powerpack_pricing_tables_enabled',
+		);
+
+		foreach ( $disabled_powerpack_features as $feature_filter_name ) {
+			add_filter( $feature_filter_name, '__return_false' );
+		}
 	}
 
 	/**
